Add files via upload
The app allows users to calculate their BMI based on their weight and height and provides feedback on whether they are underweight, normal weight, overweight, or obese. It serves as a useful tool for anyone interested in monitoring their body mass index as part of a fitness or health journey.